Add truncate and pop to ProofListIndexProxy [ECR-3669]:
Extract these operations to the interface. Move their tests
to the base test of interface operations, make some of them
parameterized.
