+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/storage/indices/ListIndex.java
@@ -90,6 +90,29 @@ public interface ListIndex<T> extends StorageIndex, Iterable<T> {
   T getLast();
 
   /**
+   * Removes the last element of the list and returns it.
+   *
+   * @return the last element of the list.
+   * @throws NoSuchElementException if the list is empty
+   * @throws IllegalStateException if this list is not valid
+   * @throws UnsupportedOperationException if this list is read-only
+   */
+  T removeLast();
+
+  /**
+   * Truncates the list, reducing its size to {@code newSize}.
+   *
+   * <p>If {@code newSize < size()}, keeps the first {@code newSize} elements, removing the rest.
+   * If {@code newSize >= size()}, has no effect.
+   *
+   * @param newSize the maximum number of elements to keep
+   * @throws IllegalArgumentException if the new size is negative
+   * @throws IllegalStateException if this list is not valid
+   * @throws UnsupportedOperationException if this list is read-only
+   */
+  void truncate(long newSize);
+
+  /**
    * Clears the list.
    *
    * @throws IllegalStateException if this list is not valid
