+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/storage/indices/BaseListIndexIntegrationTestable.java
@@ -44,6 +44,8 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 /**
  * Base class for common ListIndex tests.
@@ -230,6 +232,114 @@ abstract class BaseListIndexIntegrationTestable
       String last = l.getLast();
 
       assertThat(last, equalTo(V2));
+    });
+  }
+
+  @Test
+  void removeLastEmptyList() {
+    runTestWithView(database::createFork, (l) -> {
+      assertThrows(NoSuchElementException.class, l::removeLast);
+    });
+  }
+
+  @Test
+  void removeLastWithSnapshot() {
+    runTestWithView(database::createSnapshot, (l) -> {
+      assertThrows(UnsupportedOperationException.class, l::removeLast);
+    });
+  }
+
+  @Test
+  void removeLastSingleElementList() {
+    runTestWithView(database::createFork, (l) -> {
+      String addedElement = V1;
+      l.add(addedElement);
+      String last = l.removeLast();
+
+      assertThat(last, equalTo(addedElement));
+      assertTrue(l.isEmpty());
+    });
+  }
+
+  @Test
+  void removeLastTwoElementList() {
+    runTestWithView(database::createFork, (l) -> {
+      l.add(V1);
+      l.add(V2);
+      String last = l.removeLast();
+
+      assertThat(last, equalTo(V2));
+      assertThat(l.size(), equalTo(1L));
+      assertThat(l.get(0), equalTo(V1));
+    });
+  }
+
+  @Test
+  void truncateNonEmptyToZero() {
+    runTestWithView(database::createFork, (l) -> {
+      l.add(V1);
+      l.truncate(0);
+
+      assertTrue(l.isEmpty());
+      assertThat(l.size(), equalTo(0L));
+    });
+  }
+
+  @Test
+  void truncateToSameSize() {
+    runTestWithView(database::createFork, (l) -> {
+      l.add(V1);
+      long newSize = l.size();
+
+      l.truncate(newSize);
+
+      assertThat(l.size(), equalTo(newSize));
+    });
+  }
+
+  @Test
+  void truncateToSmallerSize() {
+    runTestWithView(database::createFork, (l) -> {
+      long newSize = 1;
+      l.add(V1);
+      l.add(V2);
+      l.truncate(newSize);
+
+      assertThat(l.size(), equalTo(newSize));
+    });
+  }
+
+  @ParameterizedTest
+  @ValueSource(longs = {3, 4, Long.MAX_VALUE})
+  void truncateToGreaterSizeHasNoEffect(long newSize) {
+    runTestWithView(database::createFork, (l) -> {
+      // Create a list of two elements
+      l.add(V1);
+      l.add(V2);
+
+      long oldSize = l.size();
+      l.truncate(newSize);
+
+      assertThat(l.size(), equalTo(oldSize));
+    });
+  }
+
+  @ParameterizedTest
+  @ValueSource(longs = {-1, -2, Long.MIN_VALUE})
+  void truncateToNegativeSizeThrows(long invalidSize) {
+    runTestWithView(database::createFork, (l) -> {
+      l.add(V1);
+
+      assertThrows(IllegalArgumentException.class,
+          () -> l.truncate(invalidSize));
+    });
+  }
+
+  @Test
+  void truncateWithSnapshot() {
+    runTestWithView(database::createSnapshot, (l) -> {
+      assertThrows(UnsupportedOperationException.class,
+          () -> l.truncate(0L));
     });
   }
